fix: prepare 2.8.1 release

+++ src/hdr/jpegxl/probe.rs
@@ -75,7 +75,7 @@ pub(crate) fn libjxl_probe_orientation_from_bytes(bytes: &[u8]) -> Option<u16> {
                         return None;
                     }
                     let info = info.assume_init();
-                    let o_ok = info.orientation as i32;
+                    let o_ok = info.orientation;
                     return ((1..=8).contains(&o_ok)).then_some(o_ok as u16);
                 }
                 libjxl_sys::JXL_DEC_SUCCESS
